feat(single): 原拓移植单文件版 + 市场详情预览可点击; fix(api): COUNT 转 int 修待复习字符串拼接
- publish-catalog: shlib 帖输出原拓坐标(pages/sents 去重数组 + g[].c 下标)
- LocalCard/目录类型增加 context; addFromZitie 落库; 学习卡自动获得'原拓'入口
- 单文件市场详情弹窗预览字点击看整页原拓(与在线版同组件)
- due-counts/decks 的 COUNT 全部 ::int; 前端求和 Number() 防字符串拼接

+++ server/routes/decks.ts
@@ -66,12 +66,12 @@ decksRouter.get('/decks', async (req: Request, res: Response) => {
             WHERE deck_id = d.id AND image_url != '' AND archived_at IS NULL ORDER BY created_at ASC LIMIT 1
         ), '') AS cover_image,
         COALESCE((
-          SELECT COUNT(*) FROM cards c
+          SELECT COUNT(*)::int FROM cards c
             LEFT JOIN user_card_progress ucp ON ucp.user_id = ${newCountP} AND ucp.card_id = c.id
             WHERE c.deck_id = d.id AND c.archived_at IS NULL AND (ucp.card_id IS NULL OR ucp.interval = 0)
         ), 0) AS new_count,
         COALESCE((
-          SELECT COUNT(*) FROM cards c
+          SELECT COUNT(*)::int FROM cards c
             INNER JOIN user_card_progress ucp ON ucp.card_id = c.id AND ucp.user_id = ${dueUserP}
             WHERE c.deck_id = d.id AND c.archived_at IS NULL AND ucp.interval > 0 AND ucp.next_review <= ${dueNowP}
         ), 0) AS due_count,
